feat(testing): Use new testing generator
